:truck: chore(api) move generate username with temporary password
