+++ api/lib/application/sco-organization-learners/sco-organization-learner-controller.js
@@ -23,26 +23,8 @@ const generateUsername = async function (request, h, dependencies = { scoOrganiz
     .code(200);
 };
 
-const generateUsernameWithTemporaryPassword = async function (
-  request,
-  h,
-  dependencies = { scoOrganizationLearnerSerializer },
-) {
-  const payload = request.payload.data.attributes;
-  const organizationId = payload['organization-id'];
-  const organizationLearnerId = payload['organization-learner-id'];
-
-  const result = await usecases.generateUsernameWithTemporaryPassword({
-    organizationLearnerId,
-    organizationId,
-  });
-
-  return h.response(dependencies.scoOrganizationLearnerSerializer.serializeCredentialsForDependent(result)).code(200);
-};
-
 const libScoOrganizationLearnerController = {
   generateUsername,
-  generateUsernameWithTemporaryPassword,
 };
 
 export { libScoOrganizationLearnerController };
