+++ api/src/prescription/organization-learner/domain/usecases/generate-username-with-temporary-password.js
@@ -2,7 +2,7 @@ import lodash from 'lodash';
 
 const { isEmpty } = lodash;
 
-import { UserNotAuthorizedToGenerateUsernamePasswordError } from '../../../src/shared/domain/errors.js';
+import { UserNotAuthorizedToGenerateUsernamePasswordError } from '../../../../shared/domain/errors.js';
 
 const generateUsernameWithTemporaryPassword = async function ({
   organizationLearnerId,
