+++ api/tests/acceptance/application/sco-organization-learners/sco-organization-learner-controller_test.js
@@ -97,81 +97,6 @@ describe('Acceptance | Controller | sco-organization-learners', function () {
     });
   });
 
-  describe('POST /api/sco-organization-learners/username-password-generation', function () {
-    let organizationId;
-    let organizationLearnerId;
-    let options;
-
-    beforeEach(async function () {
-      organizationId = databaseBuilder.factory.buildOrganization({
-        type: 'SCO',
-        isManagingStudents: true,
-      }).id;
-      const userId = databaseBuilder.factory.buildUser.withRawPassword({
-        username: null,
-      }).id;
-      databaseBuilder.factory.buildMembership({ organizationId, userId });
-      organizationLearnerId = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId,
-        userId,
-      }).id;
-
-      await databaseBuilder.commit();
-
-      options = {
-        method: 'POST',
-        url: '/api/sco-organization-learners/username-password-generation',
-        headers: generateAuthenticatedUserRequestHeaders({ userId }),
-        payload: {
-          data: {
-            attributes: {
-              'organization-id': organizationId,
-              'organization-learner-id': organizationLearnerId,
-            },
-          },
-        },
-      };
-    });
-
-    it('should return a 200 status after having successfully generated username and temporary password', async function () {
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
-
-    it('should return a 403 status when student does not belong to the same organization as organizationLearner', async function () {
-      // given
-      options.payload.data.attributes['organization-id'] = organizationId + 1;
-      options.payload.data.attributes['organization-learner-id'] = organizationLearnerId;
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
-
-    it('should return a 403 status when user does not belong to the same organization as organizationLearner', async function () {
-      // given
-      const wrongOrganization = databaseBuilder.factory.buildOrganization();
-      const organizationLearnerWithWrongOrganization = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId: wrongOrganization.id,
-      });
-      await databaseBuilder.commit();
-
-      options.payload.data.attributes['organization-learner-id'] = organizationLearnerWithWrongOrganization.id;
-      options.payload.data.attributes['organization-id'] = wrongOrganization.id;
-
-      // when
-      const response = await server.inject(options);
-
-      // then
-      expect(response.statusCode).to.equal(403);
-    });
-  });
-
   describe('POST /api/sco-organization-learners/account-recovery', function () {
     it('should return a 200 status and student information for account recovery', async function () {
       // given
