+++ api/tests/prescription/organization-learner/unit/domain/usecases/generate-username-with-temporary-password_test.js
@@ -1,6 +1,6 @@
-import { generateUsernameWithTemporaryPassword } from '../../../../lib/domain/usecases/generate-username-with-temporary-password.js';
-import { UserNotAuthorizedToGenerateUsernamePasswordError } from '../../../../src/shared/domain/errors.js';
-import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { generateUsernameWithTemporaryPassword } from '../../../../../../src/prescription/organization-learner/domain/usecases/generate-username-with-temporary-password.js';
+import { UserNotAuthorizedToGenerateUsernamePasswordError } from '../../../../../../src/shared/domain/errors.js';
+import { catchErr, domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | UseCase | generate-username-with-temporary-password', function () {
   const expectedUsername = 'john.harry0207';
